Finals_Dash/requirements.txt
streamlit
psycopg2-binary
pandas
plotly
streamlit-autorefresh
